backend/requirements.txt
langgraph
tavily-python
langchain
langchain_openai
pypandoc
pandoc
tldextract
python-dotenv
fpdf


